requirements.txt
numpy>=1.26,<2.0
pandas>=2.0,<3.0
akshare
yfinance>=0.2.30
plotly>=5.20.0,<6.0
fastapi>=0.110.0,<1.0
uvicorn[standard]>=0.29.0,<1.0
requests
pytz
vectorbt==0.26.2
numba==0.59.1
ipykernel
nbformat
pyyaml
